correção da taxa para float

+++ src/components/RowHistorico.js
@@ -3,7 +3,7 @@ import './RowHistorico.css'
 
 function Row(props) {
     const num = props.numero
-    const taxa = parseInt(props.taxa)
+    const taxa = parseFloat(props.taxa)
     return (
         <div className="row">
             <h2 id="comanda">Nº Comanda: { num in [1,2,3,4,5,6,7,8,9,10]? '0'+ props.numero: props.numero}</h2>
